+++ Changelog.md
@@ -1,3 +1,14 @@
+# v0.4.2
+
+#### Features
+
+- Remove highlights at cursor on corrections [#88](https://github.com/upfrontIO/editable.js/pull/88)
+
+#### Bugfixes
+
+- Fix spellcheck whitespace handling [#87](https://github.com/upfrontIO/editable.js/pull/87)
+
+
 # v0.4.1
 
 #### Features